+++ dan_layer/template_test_tooling/src/lib.rs
@@ -1,6 +1,5 @@
 //  Copyright 2022 The Tari Project
 //  SPDX-License-Identifier: BSD-3-Clause
-//
 
 mod mock_runtime_interface;
 pub use mock_runtime_interface::*;
