+++ dan_layer/template_test_tooling/src/template_test.rs
@@ -1,36 +1,12 @@
-//  Copyright 2022. The Tari Project
+//  Copyright 2022 The Tari Project
+//  SPDX-License-Identifier: BSD-3-Clause
 //
-//  Redistribution and use in source and binary forms, with or without modification, are permitted provided that the
-//  following conditions are met:
-//
-//  1. Redistributions of source code must retain the above copyright notice, this list of conditions and the following
-//  disclaimer.
-//
-//  2. Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the
-//  following disclaimer in the documentation and/or other materials provided with the distribution.
-//
-//  3. Neither the name of the copyright holder nor the names of its contributors may be used to endorse or promote
-//  products derived from this software without specific prior written permission.
-//
-//  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES,
-//  INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-//  DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-//  SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-//  SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
-//  WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
-//  USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 use std::{collections::HashMap, path::Path};
 
 use borsh::BorshDeserialize;
 use tari_crypto::ristretto::RistrettoSecretKey;
-use tari_template_lib::{
-    args::Arg,
-    models::{ComponentAddress, TemplateAddress},
-};
-
-use super::MockRuntimeInterface;
-use crate::{
+use tari_dan_engine::{
     crypto::create_key_pair,
     hashing::hasher,
     packager::{LoadedTemplate, Package, TemplateModuleLoader},
@@ -39,6 +15,12 @@ use crate::{
     transaction::{Instruction, TransactionBuilder, TransactionProcessor},
     wasm::{compile::compile_template, LoadedWasmTemplate},
 };
+use tari_template_lib::{
+    args::Arg,
+    models::{ComponentAddress, TemplateAddress},
+};
+
+use super::MockRuntimeInterface;
 
 pub struct TemplateTest<R> {
     package: Package,
