+++ dan_layer/template_test_tooling/tests/templates/private_function/src/lib.rs
